requirements.txt
# ./requirements.txt
openai
python-dotenv
gradio
requests




